+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/informers/cache/Store.java
@@ -105,4 +105,11 @@ public interface Store<T> {
    */
   void isPopulated(boolean isPopulated);
 
+  /**
+   * true if synced
+   */
+  default boolean hasSynced() {
+      return true;
+  }
+
 }